[New] ImGUISerializer added, initial version
ImGUISerializer is supposed to be able to generate appropriate inspection UI for the Resource that recieves it. It provides ability for the user to change the properties of the objects being inspected. A work in progress - will be used for the EntitiesWidget that will follow after this to proof it. Needs improvement to which UI widgets it's using for specific types of data it inspects.