BugFix: Fixing thread-race of accesing game state when preparing to render
Previously creation of render jobs&reading game state was done in SubmitRenderables. SubmitRenderables runs on main thread in parallel to entire engine task thread, meaning it can easily overlap with any task that updates game state. With Render task being a sync point, it's the only time to grab game state safely, so moved the relevant code to that stage.
    * Part of that fix allowed to simplify the renderable buffering into 1 extra buffer rather than RWBuffer of extras. Next step is the removal of extra buffering