Update: Provide an optional toggle to record long frame for profiling
By default set to true. I never fully exposed it, and now it'll be useful as it adds additional pressure on memory use (since it's a self-imploding situation - new long profile needs to be stored and tracked, which can worsen the perf, which can add a new long frame record).