BugFix: Dynamic resources now gain an ID if a GPUResource is created for them
This is to help avoid the pitfall of creating a custom Resource and assign it to a new GameObject frequently - since Graphics had no ID to track it with, it used to create a new GPUResource for every new assignation. This way Graphics will automatically deduplicate GPUResources for a dynamic Resource while there's a GPUResource instance around.
As an extra, checked places where checks for InvalidId were present, and tried to simplify/clean them.