BugFix: fix for missing unregistration of not-fully initialized GPUResources
Just noticed how my dynamic resource registration revealed a multithreading bug - it didn't schedule unregistrarion of GPUResources that were about to be discarded in the middle of Creation/Uploading. My comment described that an unregister command would be scheduled by the destructor, but I didn't realize it was only true for Valid GPUResources.