Update: Switching on parallel RenderJob generation for GameObjects
I ran out of ideas on how to improve the performance of RenderJob creation for GameObjects within the current design(all the additional overhead of using handles). It's about 1.2micros per render-job (on average). Also removed a profiler scope on a surprisingly small-time-of-execution TriggerCallback to reduce a bit of overhead there.