Update: UniformBuffer management has been moved out of VisualObject and into RenderPasses
This simplifies tracking of Pipeline loading state (and GPU resource creation), but mainly allows easy management of using different pipelines during rendering - previously there was no way to inject extra infromation for a VisualObject that needs to be rendered with a different Pipeline.

The tradeoff is that it's less performant, since now I need to dynamically request a sized-UBO - and that requires a bit of bookkeeping. Haven't tested the impact yet, but I don't have ideas for improvements yet anyway.