Bugfix: Fix invalid rendering of GameObjects caused by bad use of UBOs
So turns out I never used mapped UBOs correctly, which resulted in the engine overwriting data that hasn't finished being uploaded yet. I've introduced fence checking between frames to ensure that UBOs have finished being uploaded and are available. This prevented the bad rendering. The trade-off is that syncing on the fence can take 3+ ms - I'll need to experiment with finer grained fencing (per UBO).

Additionally, I've fixed a potential bug in the UniformBuffer that could cause skipping of reader head movement in a circular buffer that handles mapped regions. This could happen if 2 write operations would happen before a read operation would be issued - because bool was used, the engine would lose information that there was an extra frame of UBO available.