Update: Disable Dither in GL by default and adding an OpenGL warning to the ignore list
Dither is disabled because it raises warnings with rendering to integer framebuffers(and the engine doesn't rely on dither, as far as I see). The warning is appearing when the engine issues glClear calls without a bound Pipeline, as far as I can see. Since this is a normal situation on my side(can call glClear while the pipeline is still loading), adding it to ignore list.