Bugfix: Fixing a thread-race with Renderable allocation and access
This bug manifested as tripping up an AssertRWMutex with garbage data - cause by dereferencing a PoolPtr in the middle/right-after realloc. Originally, I've used Pool, which is not thread-safe during growth, so I've replaced it with StableVector(stronger guarantees). The crash is gone, but I've lost the ability to run parallel execution on renderables - need to investigate that next.