Clean: Simplifying I/-RenderPass machinery by removing base Process call
It forced all render passes to have to inherit from IRenderPass::Params and was generally awkward, since the IRenderPass::Params had members that might've been unnecessary. Upon closer inspection, I didn't need any of this extra cruft, so removed it.

Hoping that I'll be able to merge IRenderPass and RenderPass together so that it's more apparant how and what must be used.