Clean: Merging IRenderPass and RenderPass
There was no hard requirement to split it originally, but I decided to do it for more of "separation-of-concern", but since then the code base mutated enough that it wasn't necessary. Now, reducing this to 1 mandatory base class makes things simpler for users. Additionally, this makes it easier to refactor in the future(merge of BeginPass&Submit)

Also, did a bit of a drive by and unified parameter names for BeginPass/Submit (no more anInterface).