Update: Part one of RenderPass refactor
Moving general object rendering into the Default render pass, isolating all the job generation into the pass. The point is to simplify how rendering is organized(goal to be able to have a render pass schedule multiple stages of jobs), and to prevent outside code knowing which passes need objects to be fed into, since it's difficult to expand (example: had a DefaultRenderPass that took GameObjects, but once added IDRenderPass it was difficult to extend to support it).