Update: Part 7 of RenderPass refactor
DebugDrawers are now scheduled in DebugRenderPass(allowed to clean up Game further). Consolidated DebugRenderPass::BeginPass() and ::Submit() so that submit is no longer necessary.

Note: This introduces a bug of caching an out-of-date(1 frame) camera for debug drawing. Not fixing in this commit as next one I'll be removing Game::ScheduleRenderables, which is responsible for updating camera.