Update: RenderPasses no longer cache the RenderPassJob nor RenderContext
This removes the 1-to-1 relationship between RenderPass and RenderPassJob. I should be able to unify PaintingRenderPass and DisplayRenderPass into 1.

Also did a change to RenderContext - now need to explicitly specify how many texture slots and framebuffer draw slots need to be active. This allows to do less work in the renderer, as well as allowing to avoid bugs related default initialization of implicit array elements during aggregate initailization.