Bugfix: Improve handling of arrays of data in UBOs
While working on adding Light support, noticed that Descriptor was incorrectly calculating alignment of members, which led to invalid data uploads and reads. Fixed, and marked that I need to remove Vec3 - as I doubt I interpreted the "merging" bug in the past correctly. Instead, I'll just remove support for it as advertised on khronos wiki (inconsitent driver support).