Update: added support for directional lights
Cleaned up the fragment shader from comments, made light's color have a stronger contribution on the ambient color. Editor mode now allows to set the rotation and the type of the light. Replaced quaternion use in LightAdapter to instead use a direction vector (since that what we want in the shader).