Update: Adding the concept of a global Uniform Adapter
The goal is to introduce a mechanism that allows to update some adapters only once per frame. For now the type is implemented to track the "global" variety, but Graphics code isn't updated to utilize it, and none of the adapters use it. Need to clean up the AdapterSourceDate from camera first