New: Adding World to the engine
Entities are now stored in a world instead of in a Game. This is initial implementation - engine always has a world(needs to be optional), serialization isn't done (I want GameObjects to be saved locally to the world), no multi-world support. But next priority is to add tools to create and save worlds.