Update: NavMeshGen - improved triangle voxelization precision
After adding a couple extra triangles at origin for upcoming testing of extracting contours, I noticed that some voxels are missing. After making a number of attempts realized that I was losing precision due to doing a space-transformation round trip. I've rewritten the voxelization step to instead do all the triangle-AABB checks in voxel space, which started picking up my missing voxels.

Scaling this up, I can still see some areas that are missfiring - for example, in some cases voxels are appearing where there shouldn't be any triangles (or maybe they are touching with the voxels). Either way, I think this is good enough for me for now, as I think I'm using "bad" setup - a lot of my vertices should be close to voxel edges due to my generating the world. In user-created worlds this should be much less frequent, but I'll keep an eye on it.